// src/app/models/showcase.model.ts
import { FormControl } from '@angular/forms';

/** Showcase interface */
export interface Showcase {
  headerTitle: string;
  headerSubtitle: string;
  trophySection: string;
  trophyPage: string;
  goalSection: string;
  TOECSection: string;
  clubSignUp: string;
  userLogIn: string;
  fbUrl?: string;
  ytbUrl?: string;
  linkedInUrl?: string;
  rules?: string;
  igUrl?: string;
  phoneNumber?: string;
  emailAddress?: string;
  addressLine?: string;
  postalCode?: string;
  city?: string;
}